Basic Enemy movevement
Finished the basics, took longer than expected...

Contains: Enemy script that can set the speed(Will always move to the left as of now)

Health, that when reaching 0 destroy the object.

A string for a Tag, should be the same as objects that will stop the enemy in their tracks.