Enemy spawning  v1
Enemy lane spawning is now done.

It uses a list for the spawn positions, one can use empty objects to indicate these

Uses a list for enemy prefabs
As well as a enemy amount slider

Has a random delay for spawning, this can be set to not random by having the values be the same.

Have to manually activate the spawning by setting GameStarted bool to true.

Have made a Layer called Enemy that does not collide with itself.