Made Enemy scrub template
Doing the same as was done with the cards, made it so that the info for the script mainly is taken from a scriptableobject.
Attack position and range are yet to be added.
(Potentially move it to be centred on the enemy rather than a chosen point.

EvilBlob now represents the Enemy Test 1