READ DESCRIPTION TowerSpot Respawn
Needs added:
1 : 1 empty gameobject (Position 0,0,0)
2: On all Tower spot script holders, need a refrence to the empty game object.
Now it is ready to go

How this works. Upon calling up the Unit the spot becomes the units child, when the unit dies, it will set the child to active once more, make it the child of the empty object, then die.
The Tower spot upon being a child of the empty object will realise the mistake of having left it's parent TowerSpot and one more become a TowerSpot child, hence reseting it.