Gyroscope functionality
Have made it possible to use the gyroscope now.

Script has some demands it need, which prevents it from using gyroscope in cases where there is none. (Example pc).
Experimental scene currently hold a cube with numbered sides, these are just to show the rotation on an apk build. Tested and works. Actual functionality is the next stage.