Minigame#6
It can now rotate a wateringcan with A&D or with a phone tilt.

Setting to be ware of(everything is on the wateringcan script)

How long to water the plants for, the objects of the dead & alive plants
The stars as all minigames need.
The water effect. (Many settings in there as well.)