Boss hand Scripts made
There is a test hand scriptableobject or a new one can be made.
Has the basic stats, if more wanted. Just ask.

Currently it ignores knockback, is insta killed by piranha, and has the basics to be effected by spells(need to double check that later)

Hand uses animator like enemies for damage indication.
Needs tag & layer Enemy.

Currently not a way to confirm if there are enemies on the same lane.