Fixed OverlordHand Anim + Collision
Added the enemy damage controller to the animator(Will give the same effects enemies have when taking damage)

Added a if to units melee in case the enemy they fight are not using the basicEnemyMovement, then look for greedyOpportunity instead and use that.

Removed the OverlordHand's function to rotate upon taking damage as it would render the attack useless.