Hand updates
The hand should now use the BOOL: ObstacleInTheWay, it is public so available to be refrenced.

OnbecomeInvisible now stops the hand and makes ObstacleInTheWay true, to restart, turn it off.

Also fixed an issue where nodes did not respawn after Units died. As they were completely shut down for some reason(Not sure if that is intentional, but it prevented their return).