Fixed Property Business
It will now run a for loop of the entire list, if any of the object in the list has a parent that is a unit, then it will skip all functions and give a CD.

If it find just 1 without a unit for a parent, it will run a while loop to filter out the 1 object without it, then declare that for purchase. 

Also added a dev cheat, because the testing of these mechanics would take too long if not.
Mana now has a context menu that can be run, bu right clicking the mana script component in the mana  object. Giving you 99 mana.