Fixed Deck & Corporate sign visuals
The deck is now in its own canvas which always takes priority to be the outermost part. (Potentially add gameover + Victory here?)

Corporate sign will now become children of the overlord on spawn, as well as that they are no longer 3d object but use spriterender. Fix the fact that they covered up everything.