Overlord Hand & Corporate sign Fix
The sign now spawns with z axis of 0, no longer blocking others.

The overlord hand would stop upon not being visible by the camera. But the object no longer had a render component (mesh renderer) The child had, but the script does not care about that.
Added a 1x1 pixel as a sprite renderer. It can now be seen by the camera and stops when it leaves the view, ready to strike once more.