Card stealing animation added
When the hand leaves the screen it collides with a collider(more in the back to make sure the hand is hidden, it is then turned off.
At this moment an animation of a hand going for the deck plays, when it end on the deck it runs an event that causes a card animation to play going from the deck, and the next hand animation leaving the same way. When they both end they revert and the Original hand is turned on again with a card in hand retreating.