Save system upgraded
Now the game saves on loading a new scene.

It now stores info on if you have ever added a unit to your hand during the card reward (Outside of Tutorial)

It now stores info on if you have met an enemy spawning or not(Does not update during tutorial)

Boss is added as option, but currently does not save the info.