Moth child boss indicators
Made some functioning outlines for all beasts and Monstrosities. They all have an animation that flickers with purple(Can be edited to look better).
They all now have an outline child, that contains a sprite editor(Contains the outline image) and a animator that will run the colour change. This child is turned on when the MothBoss uses the Motherly Embrace ability on them.