Tutorial Fixes!
Added for Minigame#6 Condition if playing on android(Would be locked if not).
Made it so that the Objective text on Minigame#6 is hidden if the Speech bubble is there.

Card reward, view is no longer shuffled on #3.
Speech bubble on #2 that covers the cards is hidden after 5 sec.
Temp Fix that also hides the Piranha like the rest(Had a different Tag than the others)

Made condition if Tutorial that the Minigame timer would not run(As this causes issues in the text bubbles)

Core Game Fixes:
#1 Deck view tutorial no longer demands double output to get out of it(Button got turned of multiple time because of script)
#2 You could not win after an enemy got past you, the bot would say he would handle it, but win condition is no living enemies( He now kills them instantly(No mercy).
#3 Not all enemies would always die because of the Qucken, with bot killing them, he can now end the game when they get to close (Cards are locked during this time, not sure why). Added condition for not contradicting messages from Bot.