Map Progression Base created
The system saves the Minigame value from the MinigameSceneScript, 3 at a time checking the stage.

When the loading scene starts it will load the appropriate Minigames with images(Temp)

Day + Evening is ? for now as they are lacking good images to use. Boss is random, and thereby will need rework when it is chosen to function.