Corruption visual effects + Tutorial Song fix
The Corruption now changes color to red-ish when entering "rage mode". Rage mode is activated when  a unit is placed on the "purge slot" spawned by the "Purification" ability.

Fixed an issue where the tutorial song didn't stop playing, despite being told to stop playing in the menu. (There's now a specific code in the main menu that tells it to shut the fuck up)