Hamster Anim Fix
We explain to the hamster that he has to continue working after the dialoge with the minions.

Defaulting into prime state of Full(Sleeping) when turned on and off, the hamster has a bool saying if it is running(It will not start running is yes), this was still yes when the hamster was turned on which made it just sleep. 
The fix was to tell it that it was not running anymore.
If it is full, nothing will change(not running when full)
If it needs charge it will set it to running again whilst making it run.