Strength Weakness System
All Units and Enemies now have a damage Type, this is set in the LanePLacedUnits as a static int what the max value can be.

If the attacker has lower damageType than the attackee, then they deal 25% extra damage/( divided with 4 also a value stored in LanePlaced Units).

If the other way less, then the same but lower damage.

If none of the above apply, deal normal damage.